+++ tests/e2e/test-covr_r.R
@@ -1,0 +1,3 @@
+
+expect_no_error(rhino::covr_r())
+expect_no_error(rhino::covr_report())