+++ tests/testthat/helpers/hello.R
@@ -1,0 +1,4 @@
+#' @export
+hello <- function() {
+  "Check out Rhino docs!"
+}